Fixed RowEvaluators

+++ src/Sharp48.Solvers/Evaluators/RowEvaluation/IRowEvaluator.cs
@@ -2,6 +2,6 @@
 {
     public interface IRowEvaluator : IEvaluator
     {
-        double Evaluate(byte[] tiles);
+        double Evaluate(uint[] tiles);
     }
 }
